+++ app/Http/Controllers/ThreadsController.php
@@ -7,5 +7,20 @@ use App\Boards;
 use App\Threads;
 
 class ThreadsController extends Controller
-{
+{  
+    public function store(Request $request){
+        $this->validate($request, [
+            "board" => "required",
+            "title" => "required",
+            "content" => "required"
+        ]);
+        $board = Boards::where("slug", $request->board)->first();
+        $thread = new Threads();
+        $thread->board_id = $board->id;
+        $thread->title = $request->title;
+        $thread->content = $request->content;
+        $thread->has_file = true;
+        $thread->save();
+        return $thread;
+    }
 }