+++ src/pages/[board]/index.svelte
@@ -1,13 +1,19 @@
 <script>
     import { params, url } from "@sveltech/routify";
+    import { onMount } from "svelte";
+    import { threads, refresh } from "../../stores/stores";
     import { isChangingPage } from "@sveltech/routify";
-    const axios = require("axios");
-
+    // Grab the board slug from route parameter
     let board = $params.board;
-    let request = axios.get("/api/boards/list/?board_slug=" + board);
+    refresh($params.board);
     $: {
-        board = $params.board;
-        request = axios.get("/api/boards/list/?board_slug=" + board);
+        // console.log($isChangingPage);
+
+        // Life Cycle variable which triggers 'false' when a page has stopped transitioning
+        // a.k.a final destination is current page and has stopped
+        if (!$isChangingPage) {
+            refresh(board);
+        }
     }
 </script>
 
@@ -19,20 +25,12 @@
 </style>
 
 <div class="board_view">
-    <h1 class="is-size-1 has-text-weight-bold has-text-centered has-text-black">
-        /{board}/
-    </h1>
+    <button
+        class="button is-small"
+        on:click={refresh($params.board)}>Refresh</button>
     <div class="threads_list">
-        {#await request}
-            Loading threads...
-        {:then { data }}
-            <ul>
-                {#each data.data as t}
-                    <li><a href={`/${board}/thread/${t.id}`}>{t.title}</a></li>
-                {:else}
-                    <li>No threads found...</li>
-                {/each}
-            </ul>
-        {/await}
+        {#each $threads as b}
+            <li>{b.title} - {b.content}</li>
+        {:else}Loading...{/each}
     </div>
 </div>